Refactor unit test (#122)
* 書き込み系のテストを分離

* 分離元の方を削除

* 単体テストを実行するように変更

* Github actionsを追加

* Update dev.yml

* Update dev.yml

* Update dev.yml

* Update dev.yml

* Update .gitmodules

* Update dev.yml

* Update dev.yml

* Update dev.yml

* Update dev.yml

* Update unittest.sh

* Update dev.yml
