+++ tests/common/common.pri
@@ -2,8 +2,10 @@
 INCLUDEPATH += $$PWD
 
 HEADERS += \
+    $$PWD/unittest_common.h \
     $$PWD/webserver.h
 
 SOURCES += \
+    $$PWD/unittest_common.cpp \
     $$PWD/webserver.cpp
