+++ tests/common/unittest_common.h
@@ -1,0 +1,14 @@
+#ifndef UNITTEST_COMMON_H
+#define UNITTEST_COMMON_H
+
+#include <QString>
+#include <QJsonObject>
+#include <QHash>
+
+namespace UnitTestCommon {
+QJsonDocument loadJson(const QString &path);
+QHash<QString, QString> loadPostHash(const QString &path);
+QHash<QString, QJsonObject> loadPostExpectHash(const QString &path);
+}
+
+#endif // UNITTEST_COMMON_H
