+++ tests/hagoromo_test/tst_hagoromo_test.cpp
@@ -5,21 +5,15 @@
 #include <QFile>
 
 #include "webserver.h"
+#include "unittest_common.h"
 #include "timelinelistmodel.h"
-#include "recordoperator.h"
-#include "feedgeneratorlistmodel.h"
 #include "columnlistmodel.h"
 #include "notificationlistmodel.h"
 #include "userprofile.h"
 #include "tools/qstringex.h"
-#include "feedgeneratorlink.h"
 #include "anyprofilelistmodel.h"
-#include "accountlistmodel.h"
 #include "common.h"
 #include "postthreadlistmodel.h"
-#include "listslistmodel.h"
-#include "listitemlistmodel.h"
-#include "listfeedlistmodel.h"
 
 class hagoromo_test : public QObject
 {
@@ -34,9 +28,6 @@ private slots:
     void cleanupTestCase();
     void test_test_TimelineListModelError();
     void test_TimelineListModelFacet();
-    void test_RecordOperator();
-    void test_RecordOperator_profile();
-    void test_FeedGeneratorListModel();
     void test_ColumnListModelMove();
     void test_ColumnListModelRemove();
     void test_ColumnListModelInsertNext();
@@ -51,18 +42,10 @@ private slots:
     void test_TimelineListModel_animated_image();
     void test_NotificationListModel_warn();
     void test_TimelineListModel_next();
-    void test_FeedGeneratorLink();
     void test_AnyProfileListModel();
-    void test_AccountListModel();
     void test_TimelineListModel_text();
     void test_TimelineListModel_reply();
     void test_PostThreadListModel();
-    void test_ListsListModel();
-    void test_ListsListModel_search();
-    void test_ListsListModel_error();
-    void test_ListItemListModel();
-    void test_ListItemListModel_error();
-    void test_ListFeedListModel();
     void test_SystemTool_ImageClip();
 
 private:
@@ -70,13 +53,7 @@ private:
     quint16 m_listenPort;
     QString m_service;
 
-    void test_RecordOperatorCreateRecord(const QByteArray &body);
-    void test_putPreferences(const QString &path, const QByteArray &body);
-    void test_putRecord(const QString &path, const QByteArray &body);
     void verifyStr(const QString &expect, const QString &actual);
-    QJsonDocument loadJson(const QString &path);
-    QHash<QString, QString> loadPostHash(const QString &path);
-    QHash<QString, QJsonObject> loadPostExpectHash(const QString &path);
 };
 
 hagoromo_test::hagoromo_test()
@@ -89,24 +66,7 @@ hagoromo_test::hagoromo_test()
 
     connect(&m_mockServer, &WebServer::receivedPost,
             [=](const QHttpServerRequest &request, bool &result, QString &json) {
-                if (request.url().path() == "/response/facet/xrpc/com.atproto.repo.createRecord") {
-                    test_RecordOperatorCreateRecord(request.body());
-                    json = "{\"cid\":\"CID\",\"uri\":\"URI\"}";
-                    result = true;
-                } else if (request.url().path()
-                                   == "/response/generator/save/xrpc/app.bsky.actor.putPreferences"
-                           || request.url().path()
-                                   == "/response/generator/remove/xrpc/"
-                                      "app.bsky.actor.putPreferences") {
-                    test_putPreferences(request.url().path(), request.body());
-                    json = "{}";
-                    result = true;
-                } else if (request.url().path().endsWith("/com.atproto.repo.putRecord")) {
-                    test_putRecord(request.url().path(), request.body());
-                    json = "{}";
-                    result = true;
-                } else if (request.url().path().endsWith(
-                                   "/xrpc/com.atproto.server.refreshSession")) {
+                if (request.url().path().endsWith("/xrpc/com.atproto.server.refreshSession")) {
                     QFileInfo file_info(request.url().path());
                     QString path = ":" + file_info.filePath();
                     QFile file(path);
@@ -173,94 +133,6 @@ void hagoromo_test::test_TimelineListModelFacet()
         verifyStr(json_doc.array().at(i).toObject().value("recordText").toString(),
                   model.item(i, TimelineListModel::TimelineListModelRoles::RecordTextRole)
                           .toString());
-    }
-}
-
-void hagoromo_test::test_RecordOperator()
-{
-    RecordOperator ope;
-    ope.setAccount(m_service + "/facet", QString(), QString(), QString(), "dummy", QString());
-    QHash<QString, QString> hash = loadPostHash(":/data/com.atproto.repo.createRecord_post.expect");
-
-    QHashIterator<QString, QString> i(hash);
-    while (i.hasNext()) {
-        i.next();
-
-        ope.clear();
-        ope.setAccount(m_service + "/facet", i.key(), "handle", "email", "accessJwt", "refreshJwt");
-        ope.setText(i.value());
-
-        QSignalSpy spy(&ope, SIGNAL(finished(bool, const QString &, const QString &)));
-        ope.post();
-        spy.wait();
-        QVERIFY2(spy.count() == 1, QString("spy.count()=%1").arg(spy.count()).toUtf8());
-
-        QList<QVariant> arguments = spy.takeFirst();
-        QVERIFY(arguments.at(0).type() == QVariant::Bool);
-        QVERIFY(arguments.at(0).toBool() == true);
-    }
-}
-
-void hagoromo_test::test_RecordOperator_profile()
-{
-    RecordOperator ope;
-    ope.setAccount(m_service + "/profile/1", "did:plc:ipj5qejfoqu6eukvt72uhyit", QString(),
-                   QString(), "dummy", QString());
-
-    QSignalSpy spy(&ope, SIGNAL(finished(bool, const QString &, const QString &)));
-    ope.updateProfile("", "", "description", "display_name");
-    spy.wait();
-    QVERIFY2(spy.count() == 1, QString("spy.count()=%1").arg(spy.count()).toUtf8());
-
-    QList<QVariant> arguments = spy.takeFirst();
-    QVERIFY(arguments.at(0).type() == QVariant::Bool);
-    QVERIFY(arguments.at(0).toBool() == true);
-}
-
-void hagoromo_test::test_FeedGeneratorListModel()
-{
-    FeedGeneratorListModel model;
-    model.setAccount(m_service + "/generator", QString(), QString(), QString(), "dummy", QString());
-    model.setDisplayInterval(0);
-    {
-        QSignalSpy spy(&model, SIGNAL(runningChanged()));
-        model.getLatest();
-        spy.wait();
-        QVERIFY2(spy.count() == 2, QString("spy.count()=%1").arg(spy.count()).toUtf8());
-
-        QVERIFY(model.rowCount() == 8);
-        QVERIFY(model.item(0, FeedGeneratorListModel::SavingRole) == true);
-        QVERIFY(model.item(1, FeedGeneratorListModel::DisplayNameRole)
-                == QStringLiteral("What's Hot"));
-        QVERIFY(model.item(1, FeedGeneratorListModel::SavingRole) == true);
-        QVERIFY(model.item(2, FeedGeneratorListModel::SavingRole) == false);
-        QVERIFY(model.item(3, FeedGeneratorListModel::SavingRole) == true);
-        QVERIFY(model.item(4, FeedGeneratorListModel::SavingRole) == false);
-        QVERIFY(model.item(5, FeedGeneratorListModel::SavingRole) == false);
-        QVERIFY(model.item(6, FeedGeneratorListModel::SavingRole) == false);
-        QVERIFY(model.item(7, FeedGeneratorListModel::DisplayNameRole)
-                == QStringLiteral("mostpop"));
-        QVERIFY(model.item(7, FeedGeneratorListModel::SavingRole) == false);
-    }
-    {
-        // save
-        model.setAccount(m_service + "/generator/save", QString(), QString(), QString(), "dummy",
-                         QString());
-        QSignalSpy spy(&model, SIGNAL(runningChanged()));
-        model.saveGenerator(
-                "at://did:plc:z72i7hdynmk6r22z27h6tvur/app.bsky.feed.generator/hot-classic");
-        spy.wait(10 * 1000);
-        QVERIFY2(spy.count() == 2, QString("spy.count()=%1").arg(spy.count()).toUtf8());
-    }
-    {
-        // remove
-        model.setAccount(m_service + "/generator/remove", QString(), QString(), QString(), "dummy",
-                         QString());
-        QSignalSpy spy(&model, SIGNAL(runningChanged()));
-        model.removeGenerator(
-                "at://did:plc:z72i7hdynmk6r22z27h6tvur/app.bsky.feed.generator/with-friends");
-        spy.wait(10 * 1000);
-        QVERIFY2(spy.count() == 2, QString("spy.count()=%1").arg(spy.count()).toUtf8());
     }
 }
 
@@ -1527,34 +1399,6 @@ void hagoromo_test::test_TimelineListModel_next()
             == "bafyreiejog3yvjc2tdg4muknodbplaib2yqftukwurd4qjcnal3zdxu4ni11_13");
 }
 
-void hagoromo_test::test_FeedGeneratorLink()
-{
-    FeedGeneratorLink link;
-
-    QVERIFY(link.checkUri("https://bsky.app/profile/did:plc:hoge/feed/aaaaaaaa") == true);
-    QVERIFY(link.checkUri("https://staging.bsky.app/profile/did:plc:hoge/feed/aaaaaaaa") == false);
-    QVERIFY(link.checkUri("https://bsky.app/feeds/did:plc:hoge/feed/aaaaaaaa") == false);
-    QVERIFY(link.checkUri("https://bsky.app/profile/did:plc:hoge/feeds/aaaaaaaa") == false);
-    QVERIFY(link.checkUri("https://bsky.app/profile/did:plc:hoge/feed/") == false);
-    QVERIFY(link.checkUri("https://bsky.app/profile/handle/feed/aaaaaaaa") == false);
-
-    QVERIFY(link.convertToAtUri("https://bsky.app/profile/did:plc:hoge/feed/aaaaaaaa")
-            == "at://did:plc:hoge/app.bsky.feed.generator/aaaaaaaa");
-
-    link.setAccount(m_service + "/generator", QString(), QString(), QString(), "dummy", QString());
-    {
-        QSignalSpy spy(&link, SIGNAL(runningChanged()));
-        link.getFeedGenerator("https://bsky.app/profile/did:plc:hoge/feed/aaaaaaaa");
-        spy.wait();
-        QVERIFY2(spy.count() == 2, QString("spy.count()=%1").arg(spy.count()).toUtf8());
-    }
-
-    QVERIFY(link.avatar() == "https://cdn.bsky.social/view_avator.jpeg");
-    QVERIFY(link.displayName() == "view:displayName");
-    QVERIFY(link.creatorHandle() == "creator.bsky.social");
-    QVERIFY(link.likeCount() == 9);
-}
-
 void hagoromo_test::test_AnyProfileListModel()
 {
     AnyProfileListModel model;
@@ -1581,45 +1425,6 @@ void hagoromo_test::test_AnyProfileListModel()
     }
 
     QVERIFY(model.rowCount() == 2);
-}
-
-void hagoromo_test::test_AccountListModel()
-{
-    QString temp_path = Common::appDataFolder() + "/account.json";
-    if (QFile::exists(temp_path)) {
-        QFile::remove(temp_path);
-    }
-
-    AccountListModel model;
-
-    model.updateAccount(m_service + "/account/account1", "id1", "password1", "did:plc:account1",
-                        "account1.relog.tech", "account1@relog.tech", "accessJwt_account1",
-                        "refreshJwt_account1", true);
-    model.updateAccount(m_service + "/account/account2", "id2", "password2", "did:plc:account2",
-                        "account2.relog.tech", "account2@relog.tech", "accessJwt_account2",
-                        "refreshJwt_account2", true);
-
-    AccountListModel model2;
-    {
-        QSignalSpy spy(&model2, SIGNAL(updatedAccount(int, const QString &)));
-        model2.load();
-        spy.wait(10 * 1000);
-        spy.wait(10 * 1000);
-        QVERIFY2(spy.count() == 2, QString("spy.count()=%1").arg(spy.count()).toUtf8());
-    }
-    QVERIFY2(model2.rowCount() == 2, QString::number(model2.rowCount()).toLocal8Bit());
-    int row = 0;
-    QVERIFY2(model2.item(row, AccountListModel::DidRole).toString() == "did:plc:account1_refresh",
-             model2.item(row, AccountListModel::DidRole).toString().toLocal8Bit());
-    QVERIFY2(model2.item(row, AccountListModel::RefreshJwtRole).toString()
-                     == "refreshJwt_account1_refresh",
-             model2.item(row, AccountListModel::RefreshJwtRole).toString().toLocal8Bit());
-    row = 1;
-    QVERIFY2(model2.item(row, AccountListModel::DidRole).toString() == "did:plc:account2_refresh",
-             model2.item(row, AccountListModel::DidRole).toString().toLocal8Bit());
-    QVERIFY2(model2.item(row, AccountListModel::RefreshJwtRole).toString()
-                     == "refreshJwt_account2_refresh",
-             model2.item(row, AccountListModel::RefreshJwtRole).toString().toLocal8Bit());
 }
 
 void hagoromo_test::test_TimelineListModel_text()
@@ -1830,248 +1635,6 @@ void hagoromo_test::test_PostThreadListModel()
     QVERIFY(model.item(row, PostThreadListModel::ThreadConnectorBottomRole).toBool() == false);
 }
 
-void hagoromo_test::test_ListsListModel()
-{
-    ListsListModel model;
-
-    model.setAccount(m_service + "/lists/lists", QString(), QString(), QString(), "dummy",
-                     QString());
-
-    model.setVisibilityType(ListsListModel::VisibilityTypeAll);
-    {
-        QSignalSpy spy(&model, SIGNAL(runningChanged()));
-        model.getLatest();
-        spy.wait();
-        QVERIFY2(spy.count() == 2, QString("spy.count()=%1").arg(spy.count()).toUtf8());
-    }
-    QVERIFY2(model.rowCount() == 3,
-             QString("model.rowCount()=%1").arg(model.rowCount()).toLocal8Bit());
-    QVERIFY(model.item(0, ListsListModel::CidRole).toString()
-            == "bafyreihnheyk74x4jts23gl23icubyarggmo37xn55pop2lpystq426bqu");
-    QVERIFY(model.item(0, ListsListModel::SubscribedRole).toBool() == false);
-    QVERIFY(model.item(1, ListsListModel::CidRole).toString()
-            == "bafyreieyd765syuilkovwe3ms3cpegt7wo3xksistzy2v4xmazrwbzlwtm");
-    QVERIFY(model.item(1, ListsListModel::SubscribedRole).toBool() == true);
-    QVERIFY(model.item(2, ListsListModel::CidRole).toString()
-            == "bafyreifeiua5ltajiaad76rdfuc6c63g5xd45ysro6cjptm5enwzqpcxdy");
-    QVERIFY(model.item(2, ListsListModel::SubscribedRole).toBool() == false);
-
-    model.clear();
-    model.setVisibilityType(ListsListModel::VisibilityTypeCuration);
-    {
-        QSignalSpy spy(&model, SIGNAL(runningChanged()));
-        model.getLatest();
-        spy.wait();
-        QVERIFY2(spy.count() == 2, QString("spy.count()=%1").arg(spy.count()).toUtf8());
-    }
-    QVERIFY2(model.rowCount() == 2,
-             QString("model.rowCount()=%1").arg(model.rowCount()).toLocal8Bit());
-    QVERIFY(model.item(0, ListsListModel::CidRole).toString()
-            == "bafyreihnheyk74x4jts23gl23icubyarggmo37xn55pop2lpystq426bqu");
-    QVERIFY(model.item(0, ListsListModel::SubscribedRole).toBool() == false);
-    QVERIFY(model.item(1, ListsListModel::CidRole).toString()
-            == "bafyreifeiua5ltajiaad76rdfuc6c63g5xd45ysro6cjptm5enwzqpcxdy");
-    QVERIFY(model.item(1, ListsListModel::SubscribedRole).toBool() == false);
-
-    model.clear();
-    model.setVisibilityType(ListsListModel::VisibilityTypeModeration);
-    {
-        QSignalSpy spy(&model, SIGNAL(runningChanged()));
-        model.getLatest();
-        spy.wait();
-        QVERIFY2(spy.count() == 2, QString("spy.count()=%1").arg(spy.count()).toUtf8());
-    }
-    QVERIFY2(model.rowCount() == 1,
-             QString("model.rowCount()=%1").arg(model.rowCount()).toLocal8Bit());
-    QVERIFY(model.item(0, ListsListModel::CidRole).toString()
-            == "bafyreieyd765syuilkovwe3ms3cpegt7wo3xksistzy2v4xmazrwbzlwtm");
-    QVERIFY(model.item(0, ListsListModel::SubscribedRole).toBool() == true);
-
-    model.clear();
-    model.setVisibilityType(ListsListModel::VisibilityTypeAll);
-    {
-        QSignalSpy spy(&model, SIGNAL(runningChanged()));
-        model.getLatest();
-        spy.wait();
-        QVERIFY2(spy.count() == 2, QString("spy.count()=%1").arg(spy.count()).toUtf8());
-    }
-    {
-        QSignalSpy spy(&model, SIGNAL(runningChanged()));
-        model.getNext();
-        spy.wait();
-        QVERIFY2(spy.count() == 2, QString("spy.count()=%1").arg(spy.count()).toUtf8());
-    }
-    QVERIFY2(model.rowCount() == 6,
-             QString("model.rowCount()=%1").arg(model.rowCount()).toLocal8Bit());
-    QVERIFY(model.item(0, ListsListModel::CidRole).toString()
-            == "bafyreihnheyk74x4jts23gl23icubyarggmo37xn55pop2lpystq426bqu");
-    QVERIFY(model.item(1, ListsListModel::CidRole).toString()
-            == "bafyreieyd765syuilkovwe3ms3cpegt7wo3xksistzy2v4xmazrwbzlwtm");
-    QVERIFY(model.item(2, ListsListModel::CidRole).toString()
-            == "bafyreifeiua5ltajiaad76rdfuc6c63g5xd45ysro6cjptm5enwzqpcxdy");
-    QVERIFY(model.item(3, ListsListModel::CidRole).toString()
-            == "bafyreihnheyk74x4jts23gl23icubyarggmo37xn55pop2lpystq426bqu_next");
-    QVERIFY(model.item(4, ListsListModel::CidRole).toString()
-            == "bafyreieyd765syuilkovwe3ms3cpegt7wo3xksistzy2v4xmazrwbzlwtm_next");
-    QVERIFY(model.item(5, ListsListModel::CidRole).toString()
-            == "bafyreifeiua5ltajiaad76rdfuc6c63g5xd45ysro6cjptm5enwzqpcxdy_next");
-}
-
-void hagoromo_test::test_ListsListModel_search()
-{
-    ListsListModel model;
-
-    model.setAccount(m_service + "/lists/search", QString(), QString(), QString(), "dummy",
-                     QString());
-    model.setSearchTarget("did:plc:user_42");
-    {
-        QSignalSpy spy(&model, SIGNAL(runningChanged()));
-        model.getLatest();
-        spy.wait();
-        spy.wait();
-        spy.wait();
-        spy.wait();
-        spy.wait();
-        spy.wait(10 * 1000);
-        QVERIFY2(spy.count() == 6, QString("spy.count()=%1").arg(spy.count()).toUtf8());
-    }
-
-    QVERIFY2(model.rowCount() == 6,
-             QString("model.rowCount()=%1").arg(model.rowCount()).toLocal8Bit());
-    QVERIFY(model.item(0, ListsListModel::SearchStatusRole).toInt()
-            == ListsListModel::SearchStatusTypeNotContains);
-    QVERIFY(model.item(1, ListsListModel::SearchStatusRole).toInt()
-            == ListsListModel::SearchStatusTypeNotContains);
-    QVERIFY(model.item(2, ListsListModel::SearchStatusRole).toInt()
-            == ListsListModel::SearchStatusTypeContains);
-    QVERIFY(model.item(3, ListsListModel::SearchStatusRole).toInt()
-            == ListsListModel::SearchStatusTypeNotContains);
-    QVERIFY(model.item(4, ListsListModel::SearchStatusRole).toInt()
-            == ListsListModel::SearchStatusTypeNotContains);
-    QVERIFY(model.item(5, ListsListModel::SearchStatusRole).toInt()
-            == ListsListModel::SearchStatusTypeNotContains);
-
-    QVERIFY2(model.item(0, ListsListModel::ListItemUriRole).toString() == "",
-             model.item(0, ListsListModel::ListItemUriRole).toString().toLocal8Bit());
-    QVERIFY2(model.item(1, ListsListModel::ListItemUriRole).toString() == "",
-             model.item(1, ListsListModel::ListItemUriRole).toString().toLocal8Bit());
-    QVERIFY2(model.item(2, ListsListModel::ListItemUriRole).toString()
-                     == "at://did:plc:ipj5qejfoqu6eukvt72uhyit/app.bsky.graph.listitem/101",
-             model.item(2, ListsListModel::ListItemUriRole).toString().toLocal8Bit());
-    QVERIFY2(model.item(3, ListsListModel::ListItemUriRole).toString() == "",
-             model.item(3, ListsListModel::ListItemUriRole).toString().toLocal8Bit());
-    QVERIFY2(model.item(4, ListsListModel::ListItemUriRole).toString() == "",
-             model.item(4, ListsListModel::ListItemUriRole).toString().toLocal8Bit());
-    QVERIFY2(model.item(5, ListsListModel::ListItemUriRole).toString() == "",
-             model.item(5, ListsListModel::ListItemUriRole).toString().toLocal8Bit());
-}
-
-void hagoromo_test::test_ListsListModel_error()
-{
-    ListsListModel model;
-
-    model.setRunning(true);
-    QVERIFY(model.getLatest() == false);
-    QVERIFY(model.getNext() == false);
-
-    model.setRunning(false);
-    model.setCursor("hoge");
-    QVERIFY(model.getLatest() == false);
-    QVERIFY(model.getNext() == false);
-}
-
-void hagoromo_test::test_ListItemListModel()
-{
-    ListItemListModel model;
-
-    model.setAccount(m_service + "/lists/list", "did:plc:ipj5qejfoqu6eukvt72uhyit", QString(),
-                     QString(), "dummy", QString());
-
-    model.setUri("at://did:plc:ipj5qejfoqu6eukvt72uhyit/app.bsky.graph.list/3k7igyxfizg27");
-    {
-        QSignalSpy spy(&model, SIGNAL(runningChanged()));
-        model.getLatest();
-        spy.wait();
-        QVERIFY2(spy.count() == 2, QString("spy.count()=%1").arg(spy.count()).toUtf8());
-    }
-    QVERIFY2(model.rowCount() == 3,
-             QString("model.rowCount()=%1").arg(model.rowCount()).toLocal8Bit());
-    QVERIFY(model.item(0, ListItemListModel::DidRole).toString()
-            == "did:plc:ipj5qejfoqu6eukvt72uhyit");
-    QVERIFY(model.item(1, ListItemListModel::DidRole).toString()
-            == "did:plc:l4fsx4ujos7uw7n4ijq2ulgs");
-    QVERIFY(model.item(2, ListItemListModel::DidRole).toString()
-            == "did:plc:mqxsuw5b5rhpwo4lw6iwlid5");
-    QVERIFY(model.uri()
-            == "at://did:plc:ipj5qejfoqu6eukvt72uhyit/app.bsky.graph.list/3k7igyxfizg27");
-    QVERIFY(model.cid() == "bafyreifeiua5ltajiaad76rdfuc6c63g5xd45ysro6cjptm5enwzqpcxdy");
-    QVERIFY(model.name() == "my accounts");
-    QVERIFY(model.avatar() == "");
-    QVERIFY(model.description() == "my accounts list");
-    QVERIFY(model.subscribed() == false);
-
-    model.setUri("at://did:plc:ipj5qejfoqu6eukvt72uhyit/app.bsky.graph.list/3k7igyxfizg27");
-    {
-        QSignalSpy spy(&model, SIGNAL(runningChanged()));
-        model.getNext();
-        spy.wait();
-        QVERIFY2(spy.count() == 2, QString("spy.count()=%1").arg(spy.count()).toUtf8());
-    }
-    QVERIFY2(model.rowCount() == 6,
-             QString("model.rowCount()=%1").arg(model.rowCount()).toLocal8Bit());
-    QVERIFY(model.item(0, ListItemListModel::DidRole).toString()
-            == "did:plc:ipj5qejfoqu6eukvt72uhyit");
-    QVERIFY(model.item(1, ListItemListModel::DidRole).toString()
-            == "did:plc:l4fsx4ujos7uw7n4ijq2ulgs");
-    QVERIFY(model.item(2, ListItemListModel::DidRole).toString()
-            == "did:plc:mqxsuw5b5rhpwo4lw6iwlid5");
-    QVERIFY(model.item(3, ListItemListModel::DidRole).toString()
-            == "did:plc:ipj5qejfoqu6eukvt72uhyit_next");
-    QVERIFY(model.item(4, ListItemListModel::DidRole).toString()
-            == "did:plc:l4fsx4ujos7uw7n4ijq2ulgs_next");
-    QVERIFY(model.item(5, ListItemListModel::DidRole).toString()
-            == "did:plc:mqxsuw5b5rhpwo4lw6iwlid5_next");
-    QVERIFY(model.uri()
-            == "at://did:plc:ipj5qejfoqu6eukvt72uhyit/app.bsky.graph.list/3k7igyxfizg27");
-    QVERIFY(model.cid() == "bafyreifeiua5ltajiaad76rdfuc6c63g5xd45ysro6cjptm5enwzqpcxdy");
-    QVERIFY(model.name() == "my accounts");
-    QVERIFY(model.avatar() == "");
-    QVERIFY(model.description() == "");
-    QVERIFY(model.subscribed() == false);
-}
-
-void hagoromo_test::test_ListItemListModel_error()
-{
-    ListItemListModel model;
-
-    model.setRunning(true);
-    QVERIFY(model.getLatest() == false);
-    QVERIFY(model.getNext() == false);
-
-    model.setRunning(false);
-    model.setCursor("hoge");
-    QVERIFY(model.getLatest() == false);
-    QVERIFY(model.getNext() == false);
-}
-
-void hagoromo_test::test_ListFeedListModel()
-{
-    ListFeedListModel model;
-
-    model.setAccount(m_service + "/lists/feed/0", "did:plc:l4fsx4ujos7uw7n4ijq2ulgs", QString(),
-                     QString(), "dummy", QString());
-    model.setUri("at://did:plc:ipj5qejfoqu6eukvt72uhyit/app.bsky.graph.list/3k7igyxfizg27");
-    {
-        QSignalSpy spy(&model, SIGNAL(runningChanged()));
-        model.getLatest();
-        spy.wait();
-        QVERIFY2(spy.count() == 2, QString("spy.count()=%1").arg(spy.count()).toUtf8());
-    }
-
-    QVERIFY2(model.rowCount() == 6,
-             QString("model.rowCount()=%1").arg(model.rowCount()).toLocal8Bit());
-}
-
 void hagoromo_test::test_SystemTool_ImageClip()
 {
     SystemTool tool;
@@ -2080,98 +1643,10 @@ void hagoromo_test::test_SystemTool_ImageClip()
                                250);
 }
 
-void hagoromo_test::test_RecordOperatorCreateRecord(const QByteArray &body)
-{
-    QJsonDocument json_doc = QJsonDocument::fromJson(body);
-    QString collection = json_doc.object().value("collection").toString();
-
-    if (collection == "app.bsky.feed.post") {
-        QString did = json_doc.object().value("repo").toString();
-        QString record = json_doc.object().value("record").toObject().value("text").toString();
-
-        QHash<QString, QJsonObject> hash =
-                loadPostExpectHash(":/data/com.atproto.repo.createRecord_post.expect");
-
-        QVERIFY2(hash.contains(did), QString("Unknown test pattern: %1").arg(did).toLocal8Bit());
-        verifyStr(hash[did].value("record").toObject().value("text").toString(), record);
-        if (!hash[did].value("record").toObject().value("facets").isNull()) {
-            QVERIFY2(json_doc.object().value("record").toObject().value("facets")
-                             == hash[did].value("record").toObject().value("facets"),
-                     did.toLocal8Bit());
-        }
-    }
-}
-
-void hagoromo_test::test_putPreferences(const QString &path, const QByteArray &body)
-{
-    QJsonDocument json_doc_expect;
-    if (path.contains("/save/")) {
-        json_doc_expect = loadJson(":/data/generator/save/app.bsky.actor.putPreferences");
-    } else {
-        json_doc_expect = loadJson(":/data/generator/remove/app.bsky.actor.putPreferences");
-    }
-
-    QJsonDocument json_doc = QJsonDocument::fromJson(body);
-
-    if (json_doc_expect.object() != json_doc.object()) {
-        qDebug().noquote().nospace() << QString("\nexpect:%1\nactual:%2\n")
-                                                .arg(json_doc_expect.toJson(), json_doc.toJson());
-    }
-    QVERIFY(json_doc_expect.object() == json_doc.object());
-}
-
-void hagoromo_test::test_putRecord(const QString &path, const QByteArray &body)
-{
-    QJsonDocument json_doc_expect;
-    if (path.contains("/profile/1/")) {
-        json_doc_expect = loadJson(":/data/profile/1/com.atproto.repo.putRecord");
-    } else {
-        QVERIFY(false);
-    }
-
-    QJsonDocument json_doc = QJsonDocument::fromJson(body);
-
-    if (json_doc_expect.object() != json_doc.object()) {
-        qDebug().noquote().nospace() << QString("\nexpect:%1\nactual:%2\n")
-                                                .arg(json_doc_expect.toJson(), json_doc.toJson());
-    }
-    QVERIFY(json_doc_expect.object() == json_doc.object());
-}
-
 void hagoromo_test::verifyStr(const QString &expect, const QString &actual)
 {
     QVERIFY2(expect == actual,
              QString("\nexpect:%1\nactual:%2\n").arg(expect, actual).toLocal8Bit());
-}
-
-QJsonDocument hagoromo_test::loadJson(const QString &path)
-{
-    QFile file(path);
-    if (file.open(QFile::ReadOnly)) {
-        return QJsonDocument::fromJson(file.readAll());
-    } else {
-        return QJsonDocument();
-    }
-}
-
-QHash<QString, QString> hagoromo_test::loadPostHash(const QString &path)
-{
-    QHash<QString, QString> hash;
-    QJsonDocument json_doc = loadJson(path);
-    for (const auto &key : json_doc.object().keys()) {
-        hash[key] = json_doc.object().value(key).toObject().value("text").toString();
-    }
-    return hash;
-}
-
-QHash<QString, QJsonObject> hagoromo_test::loadPostExpectHash(const QString &path)
-{
-    QHash<QString, QJsonObject> hash;
-    QJsonDocument json_doc = loadJson(path);
-    for (const auto &key : json_doc.object().keys()) {
-        hash[key] = json_doc.object().value(key).toObject().value("expect").toObject();
-    }
-    return hash;
 }
 
 QTEST_MAIN(hagoromo_test)
