+++ tests/hagoromo_test2/data/generator/remove/app.bsky.actor.putPreferences
