+++ tests/hagoromo_test2/data/generator/save/app.bsky.actor.putPreferences
