+++ tests/hagoromo_test2/hagoromo_test2.pro
@@ -1,0 +1,18 @@
+QT += testlib httpserver gui quick
+
+CONFIG += qt console warn_on depend_includepath testcase
+CONFIG -= app_bundle
+
+TEMPLATE = app
+DEFINES += HAGOROMO_UNIT_TEST
+
+SOURCES +=  tst_hagoromo_test2.cpp
+
+include(../common/common.pri)
+include(../../lib/lib.pri)
+include(../../app/qtquick/qtquick.pri)
+include(../../openssl/openssl.pri)
+
+RESOURCES += \
+    hagoromo_test2.qrc
+
