+++ tests/hagoromo_test2/hagoromo_test2.qrc
@@ -1,0 +1,32 @@
+<RCC>
+    <qresource prefix="/">
+        <file>data/com.atproto.repo.createRecord_post.expect</file>
+        <file>response/facet/xrpc/app.bsky.actor.getProfiles</file>
+        <file>response/generator/xrpc/app.bsky.unspecced.getPopularFeedGenerators</file>
+        <file>response/generator/xrpc/app.bsky.actor.getPreferences</file>
+        <file>response/generator/save/xrpc/app.bsky.actor.getPreferences</file>
+        <file>data/generator/save/app.bsky.actor.putPreferences</file>
+        <file>data/generator/remove/app.bsky.actor.putPreferences</file>
+        <file>response/generator/remove/xrpc/app.bsky.actor.getPreferences</file>
+        <file>response/generator/xrpc/app.bsky.feed.getFeedGenerator</file>
+        <file>response/account/account1/xrpc/com.atproto.server.refreshSession</file>
+        <file>response/account/account2/xrpc/com.atproto.server.refreshSession</file>
+        <file>response/account/account1/xrpc/app.bsky.actor.getProfile</file>
+        <file>response/account/account2/xrpc/app.bsky.actor.getProfile</file>
+        <file>response/lists/lists/xrpc/app.bsky.graph.getLists</file>
+        <file>response/lists/lists/xrpc/app.bsky.graph.getLists_1</file>
+        <file>response/lists/list/xrpc/app.bsky.graph.getList_3k7igyxfizg27</file>
+        <file>response/lists/list/xrpc/app.bsky.graph.getList_3k7igyxfizg27_1</file>
+        <file>response/lists/feed/0/xrpc/app.bsky.actor.getPreferences</file>
+        <file>response/lists/feed/0/xrpc/app.bsky.feed.getListFeed</file>
+        <file>response/lists/search/xrpc/app.bsky.graph.getLists</file>
+        <file>response/lists/search/xrpc/app.bsky.graph.getLists_1</file>
+        <file>response/lists/search/xrpc/com.atproto.repo.listRecords</file>
+        <file>response/lists/search/xrpc/com.atproto.repo.listRecords_1</file>
+        <file>response/lists/search/xrpc/com.atproto.repo.listRecords_2</file>
+        <file>response/lists/search/xrpc/com.atproto.repo.listRecords_3</file>
+        <file>response/lists/search/xrpc/app.bsky.graph.getLists_2</file>
+        <file>data/profile/1/com.atproto.repo.putRecord</file>
+        <file>response/profile/1/xrpc/com.atproto.repo.getRecord</file>
+    </qresource>
+</RCC>
