+++ tests/hagoromo_test2/response/account/account2/xrpc/com.atproto.server.refreshSession
