+++ tests/hagoromo_test2/response/generator/remove/xrpc/app.bsky.actor.getPreferences
