+++ tests/hagoromo_test2/response/generator/save/xrpc/app.bsky.actor.getPreferences
