+++ tests/hagoromo_test2/response/generator/xrpc/app.bsky.actor.getPreferences
