+++ tests/hagoromo_test2/response/profile/1/xrpc/com.atproto.repo.getRecord
