+++ tests/tests.pro
@@ -3,6 +3,7 @@ TEMPLATE = subdirs
 SUBDIRS += \
     atprotocol_test \
     hagoromo_test \
+    hagoromo_test2 \
     http_test \
     search_test
